Reduced code, added tests, and updated docs.
