+++ packages/gasket-plugin-config/README.md
@@ -171,15 +171,22 @@ module.exports = {
 
 ### Config with Public config
 
-If you are not using Redux, but still need access to config values in client-side code, you can define a `public` property in your `gasket.config.js`. The config plugin will return these `public` properties to your browser, to be accessed by the `@gasket/data` plugin.
+If you are not using Redux, but still need access to config values in client-side code, you can define a `public` property in your `gasket.config.js`.
 
 ```js
 module.exports = {
   public: {
-    test1: 'config values here',
-    test2: 'config values here'
+    test1: 'config value 1 here',
+    test2: 'config value 2 here'
   }
 };
+```
+The config plugin will return these `public` properties to your browser, to be accessed by the `@gasket/data` plugin, and used like so:
+
+```js
+import gasketData from '@gasket/data';
+
+console.log(gasketData.test1); // config value 1 here
 ```
 
 ## Lifecycles
