+++ packages/gasket-plugin-config/test/middleware.test.js
@@ -1,0 +1,96 @@
+const { promisify } = require('util');
+const middleware = require('../lib/middleware');
+const { ENV_CONFIG } = require('../lib/constants');
+const sinon = require('sinon');
+
+describe('middleware', function () {
+  let gasket, mockReq, mockRes;
+
+  beforeEach(() => {
+    gasket = {
+      config: {},
+      execWaterfall: jest.fn((event, config) => Promise.resolve(config))
+    };
+
+    mockReq = sinon.stub({ mock: 'request', cookies: { market: 'de-DE' } });
+    mockRes = sinon.stub({ mock: 'response', locals: {} });
+  });
+
+  afterEach(function () {
+    sinon.restore();
+  });
+
+  it('executes before the "redux" middleware', async () => {
+    expect(middleware.timing.before).toContain('@gasket/plugin-redux');
+  });
+
+  it('sets a `config` property on the request object', async () => {
+    const mockConfig = { some: 'config' };
+    gasket[ENV_CONFIG] = mockConfig;
+    const middlewareMock = promisify(middleware.handler(gasket));
+
+    await middlewareMock(mockReq, mockRes);
+
+    expect(mockReq).toHaveProperty('config', mockConfig);
+  });
+
+  it('allows `appRequestConfig` hooks to modify the config', async () => {
+    gasket[ENV_CONFIG] = { some: 'config' };
+    gasket.execWaterfall.mockImplementation((event, config, req, res) => {
+      expect(event).toEqual('appRequestConfig');
+      expect(req).toEqual(mockReq);
+      expect(res).toEqual(mockRes);
+
+      return {
+        ...config,
+        locale: req.cookies.market
+      };
+    });
+    const middlewareMock = promisify(middleware.handler(gasket));
+
+    await middlewareMock(mockReq, mockRes);
+
+    expect(mockReq.config).toEqual({
+      some: 'config',
+      locale: 'de-DE'
+    });
+  });
+
+  it('does not swallow errors from `appRequestConfig` hooks', async () => {
+    gasket[ENV_CONFIG] = { some: 'config' };
+    gasket.execWaterfall.mockImplementation(() => {
+      return Promise.reject(new Error('Something bad'));
+    });
+    const middlewareMock = promisify(middleware.handler(gasket));
+
+    try {
+      await middlewareMock(mockReq, mockRes);
+    } catch (err) {
+      expect(err).toBeInstanceOf(Error);
+      return;
+    }
+
+    throw new Error('Middleware should have propagated an error');
+  });
+
+  it('adds public config property to locals response', async () => {
+    const mockConfig = { public: { test: 'config' } };
+    gasket[ENV_CONFIG] = mockConfig;
+    const middlewareMock = promisify(middleware.handler(gasket));
+
+    await middlewareMock(mockReq, mockRes);
+
+    expect(mockRes.locals.gasketData).toHaveProperty('config', { test: 'config' });
+  });
+
+  it('adds nothing to locals response if public config not set', async () => {
+    const mockConfig = { };
+    gasket[ENV_CONFIG] = mockConfig;
+    const middlewareMock = promisify(middleware.handler(gasket));
+
+    await middlewareMock(mockReq, mockRes);
+
+    expect(mockRes.locals.gasketData).toBeUndefined();
+  });
+
+});
